requirements.txt
absl-py==1.0.0
anytree==2.8.0
astunparse==1.6.3
Bootstrap-Flask==2.0.2
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
et-xmlfile==1.1.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
flatbuffers==2.0
gast==0.5.3
gensim==4.1.2
gevent==21.12.0
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.44.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.2
itsdangerous==2.1.1
Jinja2==3.0.3
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.0
networkx==2.7.1
nltk==3.7
numpy==1.22.2
oauthlib==3.2.0
opencv-python==4.5.5.64
openpyxl==3.0.9
opt-einsum==3.3.0
Pillow==9.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyemd==0.5.1
python-Levenshtein==0.12.2
regex==2022.3.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
six==1.16.0
smart-open==5.2.1
SQLAlchemy==1.4.32
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tqdm==4.63.0
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.13.3
WTForms==3.0.1
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
